feat: api response값 화면 세팅
- 현재 진행 중인 자서전 인터뷰 진행률 조회 화면 연결
- 생성된 소재 조회 화면 연결

Related to: #52

+++ app/composeApp/src/commonMain/kotlin/com/tdd/talktobook/data/entity/response/autobiography/CurrentInterviewProgressResponseDto.kt
@@ -6,7 +6,7 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class CurrentInterviewProgressResponseDto (
     @SerialName("progressPercentage")
-    val progressPercentage: Double = 0.0,
+    val progressPercentage: Float = 0f,
     @SerialName("status")
     val status: String = ""
 )
