+++ app/composeApp/src/commonMain/kotlin/com/tdd/talktobook/domain/entity/response/autobiography/CurrentInterviewProgressModel.kt
@@ -3,6 +3,6 @@ package com.tdd.talktobook.domain.entity.response.autobiography
 import com.tdd.talktobook.domain.entity.enums.AutobiographyStatusType
 
 data class CurrentInterviewProgressModel(
-    val progressPercentage: Double = 0.0,
+    val progressPercentage: Float = 0f,
     val status: AutobiographyStatusType = AutobiographyStatusType.EMPTY,
 )
